Move "Plugin" to main library
