+++ src/bluehawk/index.ts
@@ -2,6 +2,7 @@ export { Bluehawk } from "./bluehawk";
 export { Document } from "./Document";
 export { OnBinaryFileFunction } from "./OnBinaryFileFunction";
 export { getBluehawk } from "./getBluehawk";
+export * from "./Plugin";
 export * from "./io/System";
 export * from "./commands";
 export * from "./parser";
