+++ src/cli/cli.ts
@@ -1,24 +1,7 @@
 import * as yargs from "yargs";
-import { getBluehawk } from "../bluehawk";
-import { loadPlugins } from "./Plugin";
+import { getBluehawk, loadPlugins, commandDir } from "../bluehawk";
 import { version as yargsVersion } from "yargs/package.json";
 import { version as bluehawkVersion } from "../../package.json";
-
-export function commandDir<T>(
-  argv: yargs.Argv<T>,
-  directory: string,
-  options?: yargs.RequireDirectoryOptions
-): yargs.Argv<T> {
-  // Centralize the workaround for commandDir with TS
-  return argv.commandDir(directory, {
-    extensions: process.env.NODE_ENV === "development" ? ["js", "ts"] : ["js"],
-    exclude: /^(?:index|.*\.test)\.[jt]s$/,
-    visit(commandModule) {
-      return commandModule.default;
-    },
-    ...options,
-  });
-}
 
 export async function run(): Promise<void> {
   const preArgv = yargs.option("plugin", {
